casted comparison of double's to float to get rid of small rounding errors which caused entire texture suite to be reloaded when it was not requested.
git-svn-id: svn://fly.thruhere.net/forsaken@1460 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/title.c
@@ -14779,7 +14779,8 @@ BOOL SetGamma( SLIDER *slider )
 	
 	Gamma = ( (double)slider->value ) / 100.0F;
 
-	if ( tempgamma == Gamma )
+	// convert to float to get rid of small rounding errors
+	if ( (float)tempgamma == (float)Gamma )
 		return TRUE;
 
 	ReleaseView();
